Version with google-login

+++ project.py
@@ -3,15 +3,243 @@ app = Flask(__name__)
 
 from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Restaurant, MenuItem
+from database_setup import Base, Restaurant, MenuItem, User
+
+from flask import session as login_session
+import random, string
+
+#imports for oauth-authorisation
+from oauth2client.client import flow_from_clientsecrets
+from oauth2client.client import FlowExchangeError
+import httplib2
+import json
+from flask import make_response
+import requests
+
+#declaring ID by referencing client secrets file
+CLIENT_ID = json.loads(
+        open('client_secrets.json', 'r').read())['web']['client_id']
+
 
 
 #Connect to Database and create database session
-engine = create_engine('sqlite:///restaurantmenu.db')
+engine = create_engine('sqlite:///restaurantmenuwithusers.db')
 Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
+
+#route to the login page
+@app.route('/login')
+def showLogin():
+    state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x in xrange(32))
+    login_session['state'] = state
+    print ("The current session state is %s" %login_session['state'])
+    print (state)
+    return render_template('login.html', STATE=state)
+
+#route for login
+#@app.route('/gconnect', methods=['POST'])
+#def gconnect():
+#    if request.args.get('state') != login_session['state']:
+#        response = make_response(json.dumps('Invalid state'), 401)
+#        response.headers['Content-Type'] = 'application/json'
+#        return response
+#    code = request.data
+#    try:
+#        # upgrade authoization code into a credentials object
+#        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+#        oauth_flow.redirect_uri = 'postmessage'
+#        credentials = oauth_flow.step2_exchange(code)
+#    #wenn bei dieser Abfrage ein Fehler auftaucht
+#    except FlowExchangeError: 
+#        response = make_response(json.dumps('Failes to upgrade the authorization code.'), 401)
+#        response.headers['Content-Type'] = 'application/json'
+#        return response
+#    #check that the access token is valid
+#    access_token = credentials.access_token
+#    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s' % access_token)
+#    h = httplib2.Http()
+#    result = json.loads(h.request(url, 'GET')[1])
+#    if result.get('error') is not None: 
+#        response = make_response(json.dumps(result.get('error')), 500)
+#        response.headers['Content-Type'] = 'application/json'
+#    
+#    #verify that the access token is used for the intended user
+#    gplus_id = credentials.id_token['sub']
+#    if result['user_id'] != gplus_id: 
+#        response = make_response(
+#                json.dumps("Token's user ID doesn't match given user ID."), 401)
+#        response.headers['Content-Type'] = 'application/json'
+#        return response
+#    
+#    #verify that access token is valid for this app
+#    if result['issued_to'] != CLIENT_ID: 
+#        response = make_response(
+#                json.dumps("Token's client ID does not match"), 401)
+#        print("Token`s client ID does not match app's!")
+#        response.headers['Content-Type'] = 'application/json'
+#        return response
+#    
+#    #Check to see if user is already logged in
+#    stored_credentials = login_session.get('credentials')
+#    stored_gplus_id = login_session.get('gplus_id')
+#    if stored_credentials is not None and gplus_id == stored_gplus_id: 
+#        response = make_response(json.dumps('Current user is already connected.'), 200)
+#        response.headers['Content-Type'] = 'application/json'
+#        
+#     #store the access token in the session for later use.
+#    login_session['credentials'] = credentials
+#    login_session['gplus_id'] = gplus_id
+#
+#    #Get user-info
+#    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfor"
+#    params = {'access_token': credentials.access_token, 'alt':'json'}
+#    answer = request.get(userinfo_url, params=params)
+#    data = json.loads(answer.text)    
+#
+#    login_session['username'] = data["name"]
+#    login_session['picture'] = data["picture"]
+#    login_session['email'] = data["email"]     
+#
+#    #hier schaffe ich eine Response, die den Namen, das Bild und die E-Mail des
+#    # Nutzers kennt. 
+#
+#    output = ''
+#    output += '<h1>Welcome, '
+#    output += login_session['username']
+#
+#    output += '!</h1>'
+#    output += '<img src="'
+#    output += login_session['picture']
+#    output += ' " style = "width: 300px; height: 300px; border-radius: 150px; -webkit-border-radius: 150px; -moz-border-radius: 150px;">'
+#    flash("you are now logged in as %s"%login_session['username'])
+#    return output          
+
+@app.route('/gconnect', methods=['POST'])                       
+def gconnect():
+# Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+# Obtain authorization code
+    code = request.data
+
+    try:
+    # Upgrade the authorization code into a credentials object
+        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+                json.dumps('Failed to upgrade the authorization code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+# Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+       % access_token)
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1])
+# If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+# Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+                json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+# Verify that the access token is valid for this app.
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+                json.dumps("Token's client ID does not match app's."), 401)
+        print ("Token's client ID does not match app's.")
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    stored_credentials = login_session.get('credentials')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_credentials is not None and gplus_id == stored_gplus_id:
+        response = make_response(json.dumps('Current user is already connected.'),
+                                 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['credentials'] = credentials
+    login_session['gplus_id'] = gplus_id
+
+# Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+
+    data = json.loads(answer.text)
+
+    login_session['username'] = data['name']
+    login_session['picture'] = data['picture']
+   
+    
+    # Check ob frisch eingeloggter Nutzer schon einen User-Account bei uns hat
+    #Wenn nicht, legen wir einen an und schreiben seine Nutzer_id in die 
+    #loggin session
+    user_id = getUserID(login_session['username'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    flash("you are now logged in as %s" % login_session['username'])
+    print ("done!")
+    return output
+                 
+#DISCONNECT
+#Hier kommt der Disconnect vom Google-Account
+
+@app.route('/gdisconnect')
+def gdisconnect():
+    print(login_session)
+    credentials = login_session.get('credentials')
+    if credentials is None:
+        response = make_response(json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    #Execute HTTP GET request to revoke current token.
+    access_token = credentials.access_token
+    #url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+       % access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+
+    if result.status == 200:
+        del login_session['access_token'] 
+        del login_session['gplus_id']
+        del login_session['username']
+        del login_session['picture']
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    else:
+	
+        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
 
 
 #JSON APIs to view Restaurant Information
@@ -38,13 +266,18 @@ def restaurantsJSON():
 @app.route('/restaurant/')
 def showRestaurants():
   restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
-  return render_template('restaurants.html', restaurants = restaurants)
+  if 'username' not in login_session:
+      return render_template('publicrestaurants.html', restaurants=restaurants)
+  else: 
+      return render_template('restaurants.html', restaurants = restaurants)
 
 #Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET','POST'])
 def newRestaurant():
+  if 'username' not in login_session: 
+      return redirect('/login')
   if request.method == 'POST':
-      newRestaurant = Restaurant(name = request.form['name'])
+      newRestaurant = Restaurant(name = request.form['name']), user_id = login_session['user_id']
       session.add(newRestaurant)
       flash('New Restaurant %s Successfully Created' % newRestaurant.name)
       session.commit()
@@ -69,6 +302,8 @@ def editRestaurant(restaurant_id):
 @app.route('/restaurant/<int:restaurant_id>/delete/', methods = ['GET','POST'])
 def deleteRestaurant(restaurant_id):
   restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  if restaurantToDelete.user_id != login_session['user_id']:
+      return "<script>function myFunction() {alert('You are not authorized to delete this restaurant. Please create your own restaurant in order to delete.');}</script><body onload='myFunction()''>"
   if request.method == 'POST':
     session.delete(restaurantToDelete)
     flash('%s Successfully Deleted' % restaurantToDelete.name)
@@ -82,8 +317,12 @@ def deleteRestaurant(restaurant_id):
 @app.route('/restaurant/<int:restaurant_id>/menu/')
 def showMenu(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+    creator = getUserInfo(restaurant.user_id)
     items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
-    return render_template('menu.html', items = items, restaurant = restaurant)
+    if 'username' not in login_session or creator.id != login_session['user_id']:
+        return render_template('publicmenu.html', items = items, restaurants=restaurants, creator = creator)
+    else:
+        return render_template('menu.html', items = items, restaurant = restaurant, creator = creator)
      
 
 
@@ -92,7 +331,7 @@ def showMenu(restaurant_id):
 def newMenuItem(restaurant_id):
   restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if request.method == 'POST':
-      newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id)
+      newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id), user_id = login_session['user_id']
       session.add(newItem)
       session.commit()
       flash('New Menu %s Item Successfully Created' % (newItem.name))
@@ -137,7 +376,24 @@ def deleteMenuItem(restaurant_id,menu_id):
         return render_template('deleteMenuItem.html', item = itemToDelete)
 
 
+def createUser(login_session):
+    newUser = User(name = login_session['username'], picture = login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(name = login_session['username']).one()
+    return user.id
 
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(name = name).one()
+        return user.id
+    except: 
+        return None
+    
 
 if __name__ == '__main__':
   app.secret_key = 'super_secret_key'
